# lib/api-zod/src/generated/types/conversationInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ConversationInput {
  productId: number;
  productName: string;
  productImage: string;
  sellerEmail: string;
  sellerName?: string;
  buyerEmail: string;
  buyerName?: string;
}