# lib/api-zod/src/generated/types/messageInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MessageInput {
  senderEmail: string;
  /** @minLength 1 */
  text: string;
}