# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Order {
  id: number;
  productId: number;
  productName: string;
  productPrice: number;
  productImage: string;
  sellerEmail: string;
  buyerEmail: string;
  buyerName: string;
  buyerPhone: string;
  buyerAddress: string;
  createdAt: Date;
}