# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderInput {
  productId: number;
  productName: string;
  productPrice: number;
  productImage: string;
  sellerEmail: string;
  buyerEmail: string;
  buyerName: string;
  buyerPhone: string;
  buyerAddress: string;
}