# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductCurrency } from './productCurrency';

export interface Product {
  id: number;
  name: string;
  price: number;
  currency: ProductCurrency;
  imageUrl: string;
  images: string[];
  size: string;
  /** @nullable */
  conditionRating?: number | null;
  description: string;
  location: string;
  sellerEmail: string;
  /** @nullable */
  sellerName?: string | null;
  sold: boolean;
  pinned: boolean;
  /** @nullable */
  pinnedUntil?: Date | null;
  verified: boolean;
  views: number;
  createdAt: Date;
}