# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductInputCurrency } from './productInputCurrency';

export interface ProductInput {
  name: string;
  price: number;
  currency?: ProductInputCurrency;
  imageUrl: string;
  images?: string[];
  size: string;
  conditionRating?: number;
  description: string;
  location: string;
  sellerEmail: string;
  sellerName?: string;
}